DAOS-3159 tests: pmemobj test for checking pmemobj thread safety
PMEMobj pool operations aren't thread safe and cause failure at DAOS
service if not used without locking. This is necessarily good for
DAOS server as it locks the entire server. In order to avoid this we
need to keep this test around to check if latest releases solve this
issue. Currently this test still fails.

Signed-off-by: Vishwanath Venkatesan <vishwanath.venkatesan@intel.com>